Fixed deadlock on aio_suspend() calls
When the verification sub-process is created using clone(CLONE_VM),
calls to aio_suspend() in the sub-process that are made after
completion of the first verification job block indefinitely,
whereas this does not occur when the CLONE_VM flag is not given.
This may be due to corruption of the internal state of the
asynchronous I/O implementation, due to unsupported use with the
above clone() call.