Adopt alternate Git commit message format
This commit marks the adoption of a Git commit message subject line
format conforming to the most commonly-used standards, in which
a Git commit message's subject line must be a phrase written in
imperative (as opposed to past tense or present participle) form.
As before, articles and other unimportant words may be omitted from
subject lines, while a terminating period is always omitted and the
first letter is always capitalized (unless it is the first letter
of a proper noun whose first letter is lowercase). Also as before,
subject lines are limited to a length of 50 characters.

-------------------------------------------------------------------

Commit UUID: 54d32c78-ee95-40d9-8a85-f77dca6271b2